+++ finance/src/test/kotlin/net/corda/finance/internal/CashConfigDataFlowTest.kt
@@ -7,7 +7,7 @@ import net.corda.finance.USD
 import net.corda.testing.node.MockNetwork
 import net.corda.testing.node.MockNetworkParameters
 import net.corda.testing.node.MockNodeParameters
-import net.corda.testing.node.internal.cordappForPackages
+import net.corda.testing.node.internal.cordappWithPackages
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.After
 import org.junit.Test
@@ -21,7 +21,7 @@ class CashConfigDataFlowTest {
     @Test
     fun `issuable currencies read in from cordapp config`() {
         val node = mockNet.createNode(MockNodeParameters(additionalCordapps = listOf(
-                cordappForPackages(javaClass.packageName).withConfig(mapOf("issuableCurrencies" to listOf("EUR", "USD")))
+                cordappWithPackages(javaClass.packageName).copy(config = mapOf("issuableCurrencies" to listOf("EUR", "USD")))
         )))
         val config = node.startFlow(CashConfigDataFlow()).getOrThrow()
         assertThat(config.issuableCurrencies).containsExactly(EUR, USD)