+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/TestCordapp.kt
@@ -1,87 +1,41 @@
 package net.corda.testing.node
 
 import net.corda.core.DoNotImplement
-import net.corda.core.internal.PLATFORM_VERSION
+import net.corda.testing.driver.DriverParameters
+import net.corda.testing.driver.NodeParameters
 import net.corda.testing.node.internal.TestCordappImpl
-import net.corda.testing.node.internal.simplifyScanPackages
-import java.nio.file.Path
 
 /**
- * Represents information about a CorDapp. Used to generate CorDapp JARs in tests.
+ * Encapsulates a CorDapp that exists on the current classpath, which can be pulled in for testing. Use [TestCordapp.Factory.findCordapp]
+ * to locate an existing CorDapp.
+ *
+ * @see DriverParameters.cordappsForAllNodes
+ * @see NodeParameters.additionalCordapps
+ * @see MockNetworkParameters.cordappsForAllNodes
+ * @see MockNodeParameters.additionalCordapps
  */
 @DoNotImplement
 interface TestCordapp {
-    /** Returns the name, defaults to "test-name" if not specified. */
-    val name: String
+    /** The package used to find the CorDapp. This may not be the root package of the CorDapp. */
+    val scanPackage: String
 
-    /** Returns the title, defaults to "test-title" if not specified. */
-    val title: String
-
-    /** Returns the version string, defaults to "1.0" if not specified. */
-    val version: String
-
-    /** Returns the vendor string, defaults to "test-vendor" if not specified. */
-    val vendor: String
-
-    /** Returns the target platform version, defaults to the current platform version if not specified. */
-    val targetVersion: Int
-
-    /** Returns the config for this CorDapp, defaults to empty if not specified. */
+    /** Returns the config to be applied on this CorDapp, defaults to empty if not specified. */
     val config: Map<String, Any>
-
-    /** Returns the set of package names scanned for this test CorDapp. */
-    val packages: Set<String>
-
-    /** Returns whether the CorDapp should be jar signed. */
-    val signJar: Boolean
-
-    /** Return a copy of this [TestCordapp] but with the specified name. */
-    fun withName(name: String): TestCordapp
-
-    /** Return a copy of this [TestCordapp] but with the specified title. */
-    fun withTitle(title: String): TestCordapp
-
-    /** Return a copy of this [TestCordapp] but with the specified version string. */
-    fun withVersion(version: String): TestCordapp
-
-    /** Return a copy of this [TestCordapp] but with the specified vendor string. */
-    fun withVendor(vendor: String): TestCordapp
-
-    /** Return a copy of this [TestCordapp] but with the specified target platform version. */
-    fun withTargetVersion(targetVersion: Int): TestCordapp
-
-    /** Returns a copy of this [TestCordapp] but with the specified CorDapp config. */
-    fun withConfig(config: Map<String, Any>): TestCordapp
-
-    /** Returns a signed copy of this [TestCordapp].
-     *  Optionally can pass in the location of an existing java key store to use */
-    fun signJar(keyStorePath: Path? = null): TestCordappImpl
 
     class Factory {
         companion object {
             /**
-             * Create a [TestCordapp] object by scanning the given packages. The meta data on the CorDapp will be the
-             * default values, which can be changed with the wither methods.
+             * Scans the current classpath to find the CorDapp that contains the given package. All the CorDapp's metdata present in its
+             * MANIFEST are inherited. If more than one location containing the package is found then an exception is thrown. An exception
+             * is also thrown if no CorDapp is found.
+             *
+             * @param scanPackage The package name used to find the CorDapp. This does not need to be the root package.
+             * @param config config to be applied with the CorDapp.
              */
             @JvmStatic
-            fun fromPackages(vararg packageNames: String): TestCordapp = fromPackages(packageNames.asList())
-
-            /**
-             * Create a [TestCordapp] object by scanning the given packages. The meta data on the CorDapp will be the
-             * default values, which can be changed with the wither methods.
-             */
-            @JvmStatic
-            fun fromPackages(packageNames: Collection<String>): TestCordapp {
-                return TestCordappImpl(
-                        name = "test-name",
-                        version = "1",
-                        vendor = "test-vendor",
-                        title = "test-title",
-                        targetVersion = PLATFORM_VERSION,
-                        config = emptyMap(),
-                        packages = simplifyScanPackages(packageNames),
-                        classes = emptySet()
-                )
+            @JvmOverloads
+            fun findCordapp(scanPackage: String, config: Map<String, Any> = emptyMap()): TestCordapp {
+                return TestCordappImpl(scanPackage = scanPackage, config = config)
             }
         }
     }