+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/internal/TestCordappImpl.kt
@@ -1,35 +1,97 @@
 package net.corda.testing.node.internal
 
+import io.github.classgraph.ClassGraph
+import net.corda.core.internal.*
+import net.corda.core.utilities.contextLogger
 import net.corda.testing.node.TestCordapp
+import org.apache.commons.lang.SystemUtils
 import java.nio.file.Path
+import java.util.*
+import java.util.concurrent.ConcurrentHashMap
+import kotlin.streams.toList
 
-data class TestCordappImpl(override val name: String,
-                           override val version: String,
-                           override val vendor: String,
-                           override val title: String,
-                           override val targetVersion: Int,
-                           override val config: Map<String, Any>,
-                           override val packages: Set<String>,
-                           override val signJar: Boolean = false,
-                           val keyStorePath: Path? = null,
-                           val classes: Set<Class<*>>
-                           ) : TestCordapp {
+/**
+ * Implementation of the public [TestCordapp] API.
+ *
+ * As described in [TestCordapp.Factory.findCordapp], this represents a single CorDapp jar on the current classpath. The [scanPackage] may
+ * be for an external dependency to the project that's using this API, in which case that dependency jar is referenced as is. On the other hand,
+ * the [scanPackage] may reference a gradle CorDapp project on the local system. In this scenerio the project's "jar" task is executed to
+ * build the CorDapp jar. This allows us to inherit the CorDapp's MANIFEST information without having to do any extra processing.
+ */
+data class TestCordappImpl(override val scanPackage: String, override val config: Map<String, Any>) : TestCordappInternal {
+    override fun withoutMeta(): TestCordappImpl = copy(config = emptyMap())
 
-    override fun withName(name: String): TestCordappImpl = copy(name = name)
+    override val jarFile: Path
+        get() {
+            val jars = TestCordappImpl.findJars(scanPackage)
+            when (jars.size) {
+                0 -> throw IllegalArgumentException("Package $scanPackage does not exist")
+                1 -> return jars.first()
+                else -> throw IllegalArgumentException("More than one jar found containing package $scanPackage: $jars")
+            }
+        }
 
-    override fun withVersion(version: String): TestCordappImpl = copy(version = version)
+    companion object {
+        private val packageToRootPaths = ConcurrentHashMap<String, Set<Path>>()
+        private val projectRootToBuiltJar = ConcurrentHashMap<Path, Path>()
+        private val log = contextLogger()
 
-    override fun withVendor(vendor: String): TestCordappImpl = copy(vendor = vendor)
+        fun findJars(scanPackage: String): Set<Path> {
+            val rootPaths = findRootPaths(scanPackage)
+            return if (rootPaths.all { it.toString().endsWith(".jar") }) {
+                // We don't need to do anything more if all the root paths are jars
+                rootPaths
+            } else {
+                // Otherwise we need to build those paths which are local projects and extract the built jar from them
+                rootPaths.mapTo(HashSet()) { if (it.toString().endsWith(".jar")) it else buildCordappJar(it) }
+            }
+        }
 
-    override fun withTitle(title: String): TestCordappImpl = copy(title = title)
+        private fun findRootPaths(scanPackage: String): Set<Path> {
+            return packageToRootPaths.computeIfAbsent(scanPackage) {
+                ClassGraph()
+                        .whitelistPackages(scanPackage)
+                        .scan()
+                        .use { it.allResources }
+                        .asSequence()
+                        .map { it.classpathElementURL.toPath() }
+                        .filterNot { it.toString().endsWith("-tests.jar") }
+                        .map { if (it.toString().endsWith(".jar")) it else findProjectRoot(it) }
+                        .toSet()
+            }
+        }
 
-    override fun withTargetVersion(targetVersion: Int): TestCordappImpl = copy(targetVersion = targetVersion)
+        private fun findProjectRoot(path: Path): Path {
+            var current = path
+            while (true) {
+                if ((current / "build.gradle").exists()) {
+                    return current
+                }
+                current = current.parent
+            }
+        }
 
-    override fun withConfig(config: Map<String, Any>): TestCordappImpl = copy(config = config)
+        private fun buildCordappJar(projectRoot: Path): Path {
+            return projectRootToBuiltJar.computeIfAbsent(projectRoot) {
+                val gradlew = findGradlewDir(projectRoot) / (if (SystemUtils.IS_OS_WINDOWS) "gradlew.bat" else "gradlew")
+                val libs = projectRoot / "build" / "libs"
+                libs.deleteRecursively()
+                log.info("Generating CorDapp jar from local project in $projectRoot ...")
+                val exitCode = ProcessBuilder(gradlew.toString(), "jar").directory(projectRoot.toFile()).inheritIO().start().waitFor()
+                check(exitCode == 0) { "Unable to generate CorDapp jar from local project in $projectRoot ($exitCode)" }
+                val jars = libs.list { it.filter { it.toString().endsWith(".jar") }.toList() }
+                checkNotNull(jars.singleOrNull()) { "Expecting a single built jar in $libs, but instead got $jars" }
+            }
+        }
 
-    override fun signJar(keyStorePath: Path?): TestCordappImpl = copy(signJar = true, keyStorePath = keyStorePath)
-
-    fun withClasses(vararg classes: Class<*>): TestCordappImpl {
-        return copy(classes = classes.filter { clazz -> packages.none { clazz.name.startsWith("$it.") } }.toSet())
+        private fun findGradlewDir(path: Path): Path {
+            var current = path
+            while (true) {
+                if ((current / "gradlew").exists() && (current / "gradlew.bat").exists()) {
+                    return current
+                }
+                current = current.parent
+            }
+        }
     }
 }